New get_timestamp() API that separates error code and time value
The ARM toolchain's time_t is unsigned, so we can't return
a negative time_t to indicate an error - it gets casted to
a very large positive time.

This commit adds a new overload

  int NTPClient::get_timestamp(time_t &timestamp, int timeout);

which returns 0 on success and a negative value on failure. On
success, the time is written to the timestamp variable. This
deprecates the old NTPClient::get_timestamp().